add ProductService, initalize table, search funcion

+++ src/main/java/com/quickmanager/model/SanPham.java
@@ -1,5 +1,6 @@
 package com.quickmanager.model;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 public class SanPham {
@@ -7,8 +8,8 @@ public class SanPham {
     private int maSanPham;
     private String tenSanPham;
     private int maDanhMuc;
-    private double giaNhap;
-    private double giaBan;
+    private BigDecimal giaNhap;
+    private BigDecimal giaBan;
     private int soLuongTon;
     private String donViTinh;
     private Date ngaySanXuat;
@@ -42,19 +43,19 @@ public class SanPham {
         this.maDanhMuc = maDanhMuc;
     }
 
-    public double getGiaNhap() {
+    public BigDecimal getGiaNhap() {
         return giaNhap;
     }
 
-    public void setGiaNhap(double giaNhap) {
+    public void setGiaNhap(BigDecimal giaNhap) {
         this.giaNhap = giaNhap;
     }
 
-    public double getGiaBan() {
+    public BigDecimal getGiaBan() {
         return giaBan;
     }
 
-    public void setGiaBan(double giaBan) {
+    public void setGiaBan(BigDecimal giaBan) {
         this.giaBan = giaBan;
     }
